Trying to fix client/server to receiver and sender
The use of the words client and server is misleading when we think of upload. Whatever we think of the possibility of using this method towards the server, I think we should anyway refer to the data sender and receiver where we can.
There are some cases (i.e. setup and signalling) where it still might make sense to talk of client and server, and I tried to leave these - this will need some careful checking. Please look