Suggestions to generalise beyond quic
We have been asked to generalise beyond QUIC, none-the-less this is something that could alreday be done in TCP, and for which we can make concrete suggestions for QUIC, so let's not do a major disruptive change unless we know what is needed. Instead, I  propose removing the direct dependency on QUIC and a modification to title and abstract to reflect this.